feat(highlights-events): adding events and roadblock color fix
