+++ src/components/Page/Team/Highlights/highlight-section.tsx
@@ -1,13 +1,8 @@
 import { Grid, GridProps, Text, VStack } from '@chakra-ui/react'
-import React, { useCallback } from 'react'
-import { useIntl } from 'react-intl'
-import { useSetRecoilState } from 'recoil'
+import React from 'react'
 
-import { configHighlightModal } from 'src/state/recoil/team/highlight/is-open-highlight-modal'
-
-import messages from './messages'
+import { HighLightCardComponent } from './highlight-card'
 import { CARD_TYPES } from './utils/card-types'
-import { highlightCardTheme } from './utils/theme'
 
 export type HighlightCard = {
   type: CARD_TYPES
@@ -22,36 +17,6 @@ interface HighlightSectionProperties {
 }
 
 const HighlightSection = ({ title, data, gridTemplate }: HighlightSectionProperties) => {
-  const intl = useIntl()
-  const setHighlightModalConfig = useSetRecoilState(configHighlightModal)
-
-  const handleOpenModal = useCallback(
-    (item: HighlightCard) => {
-      setHighlightModalConfig({
-        isOpen: true,
-        type: item.type,
-        usersIds: item.usersIds ?? undefined,
-      })
-    },
-    [setHighlightModalConfig],
-  )
-
-  const cardTitle = new Map([
-    [CARD_TYPES.FEELING, intl.formatMessage(messages.cardTitle, { type: CARD_TYPES.FEELING })],
-    [
-      CARD_TYPES.PRODUCTIVITY,
-      intl.formatMessage(messages.cardTitle, { type: CARD_TYPES.PRODUCTIVITY }),
-    ],
-    [CARD_TYPES.ROADBLOCK, intl.formatMessage(messages.cardTitle, { type: CARD_TYPES.ROADBLOCK })],
-    [CARD_TYPES.CHECKIN, intl.formatMessage(messages.cardTitle, { type: CARD_TYPES.CHECKIN })],
-    [
-      CARD_TYPES.CONFIDENCE,
-      intl.formatMessage(messages.cardTitle, { type: CARD_TYPES.CONFIDENCE }),
-    ],
-    [CARD_TYPES.KRMEMBERS, intl.formatMessage(messages.cardTitle, { type: CARD_TYPES.KRMEMBERS })],
-    [CARD_TYPES.BARRIER, intl.formatMessage(messages.cardTitle, { type: CARD_TYPES.BARRIER })],
-  ])
-
   return (
     <VStack>
       <Text w="100%" fontSize={14} color="new-gray.800" fontWeight="medium">
@@ -59,47 +24,7 @@ const HighlightSection = ({ title, data, gridTemplate }: HighlightSectionPropert
       </Text>
       <Grid gap={3} w="100%" templateColumns={gridTemplate}>
         {data?.map((item) => (
-          <VStack
-            key={item.type}
-            cursor={item.quantity >= 1 ? 'pointer' : 'default'}
-            textAlign="center"
-            alignItems="center"
-            justifyContent="center"
-            bg={highlightCardTheme[item.type].bg}
-            height="80px"
-            lineHeight="100%"
-            color={highlightCardTheme[item.type].color}
-            transition="background-color .4s"
-            _hover={{
-              bg: highlightCardTheme[item.type].hover,
-            }}
-            borderRadius={9}
-            onClick={() => (item.quantity > 0 ? handleOpenModal(item) : undefined)}
-          >
-            <Text
-              flex={2}
-              display="flex"
-              justifyContent="center"
-              alignItems="center"
-              fontSize={11}
-              fontWeight="bold"
-              maxW={120}
-              wordBreak="break-word"
-            >
-              {cardTitle.get(item.type)?.toUpperCase()}
-            </Text>
-            <Text
-              flex={1}
-              display="flex"
-              justifyContent="center"
-              alignItems="start"
-              lineHeight={0}
-              fontSize={28}
-              fontWeight="medium"
-            >
-              {item.quantity}
-            </Text>
-          </VStack>
+          <HighLightCardComponent key={item.type} item={item} />
         ))}
       </Grid>
     </VStack>
