+++ src/components/Routine/RetrospectiveTab/Answers/utils/contants.ts
@@ -1,5 +1,5 @@
 export const themeColor = (type: string) => {
   if (type === 'emoji_scale') return 'yellow.700'
   if (type === 'value_range') return 'brand.500'
-  if (type === 'road_block') return 'purple.500'
+  if (type === 'road_block') return 'pink.500'
 }
