+++ src/state/hooks/useEvent/events/team-highlight-click-data.ts
@@ -1,0 +1,3 @@
+import { BaseEventData } from './base-event'
+
+export interface TeamHighlightClickData extends BaseEventData {}
